Utilize code separation induced by extending com_link_manager
the general link manager is now an endpoint and provides override behavior for methods in com_link_manager. Given the message targets specific to copious-transitions are not in the _instance_targets table, the custom method calls for bd, transition_engine, etc. will be used in the general link manager case.